Copy libeay32.dll and ssleay32.dll to PluggableTransports directory.
M2Crypto needs to be able to dynamically link with the OpenSSL DLLs. The
DLLs are present one directory up (in the same directory as tor.exe),
but that directory isn't in the DLL search path. Putting the DLLs in the
same directory as the executables is enough to put them on the search
path.