Install go.
We use a small patch in order to enable cgo when cross-compiling. One of
the things that needs cgo is the crypto/x509 library, which is needed on
OS X to read system CA certificates. The patch and workaround are from
here:
https://code.google.com/p/go/issues/detail?id=4714#c7

The error you get if cgo is not enabled on OS X is
"x509: failed to load system roots and no roots provided".

Also enable cgo for Windows, even though it's not needed to read the
certificate store on Windows.