Make an iconless shallow copy of TorBrowser.app.
This is an implementation of the icon-hiding idea from
https://trac.torproject.org/projects/tor/ticket/11429#comment:5. The
Info.plist file enables us to control whether an icon appears, but I
could find no way to override the values we need at run time. So instead
we create a copy of the original bundle directory, with the only
difference being the contents of Info.plist. (The "copy" is actually a
few symlinks, so it doesn't take much space.)

The shallow directory copy is in a directory named
TorBrowser.app.meek-http-helper. I made the name not end in ".app" so
that you can't open it as a bundle by accident from the Finder. I put a
README in the bundle directory explaining why the directory exists.

There's a new build helper, background-plist.py, which deletes the
LSUIElement key (if present) and sets LSBackgroundOnly=true. This is
more robust than a sed script that inserts a LSBackgroundOnly key
somewhere in the XML, because the same key could be already set, or
LSUIElement have a contradictory value, elsewhere in the file.

I chose to set the key LSBackgroundOnly, rather than the LSUIElement
we've discussed so far, though both of them work in terms of hiding the
icon. I made this decision based only on the documentation for the two
keys, which makes LSBackgroundOnly sound more background-y.

https://developer.apple.com/library/mac/documentation/general/Reference/InfoPlistKeyReference/Articles/LaunchServicesKeys.html#//apple_ref/doc/uid/20001431-108256
	LSUIElement specifies whether the app runs as an agent app. If
	this key is set to “1”, Launch Services runs the app as an agent
	app. Agent apps do not appear in the Dock or in the Force Quit
	window. Although they typically run as background apps, they can
	come to the foreground to present a user interface if desired. A
	click on a window belonging to an agent app brings that app
	forward to handle events.

https://developer.apple.com/library/mac/documentation/general/Reference/InfoPlistKeyReference/Articles/LaunchServicesKeys.html#//apple_ref/doc/uid/20001431-109268
	LSBackgroundOnly specifies whether this app runs only in the
	background. If this key exists and is set to “1”, Launch
	Services runs the app in the background only. You can use this
	key to create faceless background apps. You should also use this
	key if your app uses higher-level frameworks that connect to the
	window server, but are not intended to be visible to users.
	Background apps must be compiled as Mach-O executables.