Bug 13857: allow updating responses for a specific channel
The update_responses command can now receive a channel name as argument.
In that case, only the xml responses for the selected channel(s) are
updated. The files are now located in directory htdocfiles/$channel and
each channel has its own .htaccess file in that directory.

Makefile rules have been added to update the different channels.