Add a command to check an update_responses deployement
usage: check_update_responses_deployement <base_url> [channels...]

example:
  $ ./check_update_responses_deployement \
        https://dist.torproject.org/torbrowser/update_2 release

This command will read the config.yml file, then try to load different
URLs to check that it returns correct xml with the expected version,
incremental update files for the versions that should have them, or no
update when already using the current version.

It does not currently check that the checksum of the files is correct.