Bug 16860: Reuse favorite Tor Browser icon
Adding "StartupWMClass=Tor Browser" makes sure that the favorite
Tor Browser icon is reused after Tor Browser got registered. If this
property is missing one gets two icons: oe for the favorite and a new
icon for the running instance. This impacts at least Gnome3.

Thanks to a cypherpunk for the proposed fix.